Replace calls to SDL_GetMouseState() with Action methods where button number matters; implement more needed support in Action

+++ src/Engine/InteractiveSurface.h
@@ -41,7 +41,7 @@ class InteractiveSurface : public Surface
 private:
 	static const int NUM_BUTTONS = 7;
 	static const SDLKey SDLK_ANY;
-	Uint8 _buttonsPressed;
+	Uint32 _buttonsPressed;
 protected:
 	std::map<Uint8, ActionHandler> _click, _press, _release;
 	ActionHandler _in, _over, _out;
